updated namespaces and removed privacy action from controller
