+++ DndOverlordHomePage/Models/Item.cs
@@ -1,4 +1,4 @@
-namespace DndOverlord;
+namespace DndOverlordHomePage.Models;
 
 public class Item
 {
