+++ DndOverlordHomePage/Models/Player.cs
@@ -1,4 +1,4 @@
-namespace DndOverlord;
+namespace DndOverlordHomePage.Models;
 
 
 public class Player : Creature
