+++ DndOverlordHomePage/Models/Spell.cs
@@ -1,4 +1,4 @@
-namespace DndOverlord;
+namespace DndOverlordHomePage.Models;
 
 public class Spell : Weapon
 {
