+++ DndOverlordHomePage/Views/Home/Index.cshtml
@@ -4,5 +4,5 @@
 
 <div class="text-center">
     <h1 class="display-4">Welcome</h1>
-    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
+    <p> Web application used to do cool things</p>
 </div>
